no message
git-svn-id: https://svn.grrrr.org/ext/trunk@221 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flclass.h
@@ -327,7 +327,7 @@ public:
 	static void SetString(t_atom &a,const char *c) { SetSymbol(a,gensym(const_cast<char *>(c))); }
 
 	static bool CanbeInt(const t_atom &a) { return IsFloat(a) || IsInt(a); }
-	static float GetAInt(const t_atom &a) { return GetAFlint(a); }
+	static int GetAInt(const t_atom &a) { return (int)GetAFlint(a); }
 
 	static bool IsFlint(const t_atom &a) { return IsFloat(a) || IsInt(a); }
 	static float GetAFlint(const t_atom &a) { return IsFloat(a)?GetFloat(a):(IsInt(a)?GetInt(a):0); }
